backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
python-dotenv
Pillow
psycopg2-binary
django-storages
boto3
mercadopago
python-decouple
